+++ TMessagesProj/src/main/kotlin/xyz/nextalone/nagram/NaConfig.kt
@@ -1332,6 +1332,12 @@ object NaConfig {
             ConfigItem.configTypeBool,
             false
         )
+    val whiteSendButton =
+        addConfig(
+            "WhiteSendButton",
+            ConfigItem.configTypeBool,
+            true
+        )
     val ignoreUnreadCount =
         addConfig(
             "IgnoreUnreadCount",